guest-app/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pyjwt==2.8.0
bcrypt==4.1.2
pymongo==4.6.0
